Deconvolve the pure tumour methylation rate from bisulfite sequencing data

get_pure_tumour_methylation

get_pure_tumour_methylation(
  patient_id,
  sample_id,
  sex,
  normal_infiltrates_proxy_id,
  path,
  path_to_CAMDAC,
  build,
  n_cores,
  reseg = FALSE
)

Arguments

patient_id

Character variable containting the patient id number

sample_id

Character variable with the (control or tumour) sample_id

sex

Character variable with the patient expressed as "XX" for female or "XY" for male.

normal_infiltrates_proxy_id,

Sample ID of the matched normal control

path

Character path variable pointing to the desired working directory. This is where the output will be stored and should be constant for all CAMDAC functions.

path_to_CAMDAC

Character variable containting the path to the CAMDAC directory including dir name (e.g. "/path/to/CAMDAC/").

build

Character variable corresponding to the reference genome used for alignment. CAMDAC is compatible with "hg19", "hg38", "GRCH37","GRCH38".

n_cores

Numerical value correspdonding to the number of cores for parallel processing

reseg

Logical value should be set to FALSE. Multi-sample re-segmentation of the copy number profiles will be available in future versions of CAMDAC.

Note: #' Annotation include: CGI (including shore and shelves) gene body (intragenic, 5UTR, 3UTR, intron, exon) promoter (2kb upstream and 500 downstream any UCSC annotated gene) enhancer (vista and FANTOM5 annotation)

Value

CAMDAC purified tumour methylation rates