Set CAMDAC configuration

Set CAMDAC configuration

CamConfig(
  outdir,
  bsseq,
  lib,
  build,
  n_cores = 1,
  regions = NULL,
  refs = NULL,
  n_seg_split = 50,
  min_mapq = 1,
  min_cov = 3,
  overwrite = FALSE,
  cna_caller = "battenberg",
  cna_settings = NULL
)

Arguments

outdir

A path to save CAMDAC results. The results folder structure follows the format PATIENT/DATASET/SAMPLE/.

bsseq

Bisulfite sequencing platform. Choose between "wgbs" or "rrbs".

lib

Bisulfite sequencing library. Choose "pe" for paired end, "se" for single end.

build

Reference genome build. Choose "hg38" or "hg19".

n_cores

Number of cores to process CAMDAC data in parallel wherever possible.

regions

A BED file with regions to restrict the analysis to

refs

Path to CAMDAC reference files. If this is not given, CAMDAC searches the environment variable CAMDAC_PIPELINE_FILES. If this is not set, CAMDAC searches recursively in the current working directory.

min_mapq

Minimum mapping quality filter used in cmain_allele_counts().

min_cov

Minimum coverage filter for: DNA methylation, Normal SNP selection.

overwrite

Config to overwrite files if they already exist.

cna_caller

The CNA caller to use. "ascat" or "battenberg". Default is "battenberg"

cna_settings

A list of settings to pass to the CNA caller. rho, psi, java, beaglemaxmem